src/main/java/com/bridgelabz/hotelreservationsystem/HotelReservation/Hotel.java
package com.bridgelabz.hotelreservationsystem.HotelReservation;

import java.time.LocalDate;
import java.time.Period;
import java.time.format.DateTimeFormatter;

public class Hotel {
	private String hotelName;
	private int rate;
	private int weekdayRate, weekendRate;
	private String startDate, endDate;

	public Hotel(String hotelName, int weekdayRate, int weekendRate) {
		this.hotelName = hotelName;
		this.weekdayRate = weekdayRate;
		this.weekendRate = weekendRate;

	}

	public Hotel(String hotelName, int rate, String startDate, String endDate) {
		this.hotelName = hotelName;
		this.rate = rate;
		this.startDate = startDate;
		this.endDate = endDate;
	}

	public int getWeekdayRate() {
		return weekdayRate;
	}

	public int getWeekendRate() {
		return weekendRate;
	}

	public String getHotelName() {
		return hotelName;
	}

	public int getRate() {
		return rate;
	}

	public int calculateNumberOfDays(String startDate, String endDate) {
		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("ddMMMyyyy");
		LocalDate localStartDate = LocalDate.parse(startDate, formatter);
		LocalDate localEndDate = LocalDate.parse(endDate, formatter);
		return Period.between(localStartDate, localEndDate).getDays();
	}

	public int calculatePrice() {
		return calculateNumberOfDays(startDate, endDate) * rate;
	}
}
